Complete
all web page are done
it project 


1.1  INTRODUCTION
The   main  objective  of  Skyline  ERP  is  to  make  easy  for  faculty’s as well  as students  to  coordinate  with  each  other .It  acts  as  a  mediator  between  the  faculty and  students. Faculty  can  update  any  information and  can  chat  with  the  registered student. It  also  view  the  students  updates.
For  the  students  also  there  are  various  modules  helping  them  to  upload documents  and  chat  with  the  registered  faculty. Every  students  and faculties  has  to  login  with their  User  Id  and  Password.
There  are  several  modules  that  work  once  when  the  faculties  or  students  is  logged in. They  can choose  any  of  the  component  displayed and  can  avail  the  services according  to  their  need.
User  Id  and  Password  is  given  to  each  faculty  and  students  so  that messaging system  can  recognizes  them  as  user. Only  users  which  is  registered  in  this  can  get the  services  through  this.
This  web  based  system  will  provide  better  prospective  for  the enhancement  of organisation  regarding  to  quality  and  transparency.
Obtaining  a User  Id  and  Password.  The  User  Id  and  Password  being unique  and  auto  generated.  After  they  get  there  registration  confirmed, he/she  can  avail  any service  according  to  requirements.
The project allows Hod to schedule a test specifying time and set of questions for the test. It would enable educational institutes to perform technical tests, aptitude tests and quiz. The result of the test would be immediately available to the student. Further the student can also check the wrong answers and their solutions after the test. The Hod can also upload related materials on the server which will be made available to the students for download. This project would be helpful for creating practice tests, say for educational institutes.
This project would be very useful for educational institutes where regular evaluation of students’ is required. Further it can also be useful for anyone who requires feedback based on objective type responses. Apart from this examination aspect, this project would be very useful for the placement department of the institutes to efficiently handle the student data in database that at present is maintained in the form of excel sheets manually.
Since this software is being designed to serve as one point solution to online examinations and online record maintenance, this software will also help institutes to maintain the records of students  related activities like filtering of student data etc.